# digest: use github.com/minio/sha256-simd