--- a/pkg/hasher/sha256.go
+++ b/pkg/hasher/sha256.go
@@ -14,30 +14,15 @@
    limitations under the License.
 */
 
-package main
+package hasher
 
 import (
-	"crypto"
-	"fmt"
-	"os"
+	"hash"
 
-	"github.com/containerd/containerd/cmd/containerd/command"
-	"github.com/containerd/containerd/pkg/hasher"
-	"github.com/containerd/containerd/pkg/seed"
-
-	_ "github.com/containerd/containerd/cmd/containerd/builtins"
-	_ "github.com/containerd/containerd/pkg/cri"
+	sha256simd "github.com/minio/sha256-simd"
 )
 
-func init() {
-	seed.WithTimeAndRand()
-	crypto.RegisterHash(crypto.SHA256, hasher.NewSHA256)
-}
-
-func main() {
-	app := command.App()
-	if err := app.Run(os.Args); err != nil {
-		fmt.Fprintf(os.Stderr, "containerd: %s\n", err)
-		os.Exit(1)
-	}
+func NewSHA256() hash.Hash {
+	// An equivalent of sha256-simd is expected to be merged in Go 1.20 (1.21?): https://go-review.googlesource.com/c/go/+/408795/1
+	return sha256simd.New()
 }