https://github.com/minio/sha256-simd simdizes sha256 computation on amd64 and arm64.

Micro benchmark: `ctr images pull docker.io/library/fedora:latest` on `Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz` (with `sha_ni`), 4 cores

Before:
```
9.110s
9.172s
9.390s
```

After:
```
8.609s
8.321s
8.654s
```