@dmcgowan 2022-11-29
Hopefully this is could be a temporary change and Go can find an upstream solution. The performance improvement is significant enough to pull in though.