@thaJeztah 2022-11-29
> Hopefully this is could be a temporary change and Go can find an upstream solution. The performance improvement is significant enough to pull in though.

Looks like it was merged in Go master; https://github.com/golang/go/issues/53084 - see https://github.com/golang/go/commit/f1b1557cf3c9ca5e3def849656f79f39041d0055 (not sure if more is needed after that>), which would mean this will be in go1.20?

If so, should we add a `TODO` comment to remove it after we switched to go1.20?