@AkihiroSuda 2022-11-30
> Looks like it was merged in Go master; [golang/go#53084](https://github.com/golang/go/issues/53084) - see [golang/go@f1b1557](https://github.com/golang/go/commit/f1b1557cf3c9ca5e3def849656f79f39041d0055) (not sure if more is needed after that>), which would mean this will be in go1.20?

`crypto/sha256: add sha-ni implementation` ( https://go-review.googlesource.com/c/go/+/408795/1 ) doesn't seem merged.
Not sure it will be merged before Go 1.20.