@AkihiroSuda 2022-11-30
Added a code comment

```go
func NewSHA256() hash.Hash {
        // An equivalent of sha256-simd is expected to be merged in Go 1.20 (1.21?): https://go-review.googlesource.com/c/go/+/408795/1
        return sha256simd.New()
}
```