@AkihiroSuda 2022-12-01
> @AkihiroSuda Would it be relevant to have this new computation in BuildKit too?

Yes, I guess we can safely have this in BuildKit (and Moby, etc.)