@cpuguy83 2022-12-01
This is pretty cool, but I wonder if this should be opt-in given that the performance advantage is fairly negligible unless you are doing a lot digesting.
This could even be a plugin that is disabled by default where the plugin init would register the hasher.