@AkihiroSuda 2022-12-02
> opt-in

Any drawback to enable SIMD by default?

> This could even be a plugin that is disabled by default where the plugin init would register the hasher.

I'm not sure. The plugin will be NOP when https://go-review.googlesource.com/c/go/+/408795/1 is merged.