@samuelkarp 2022-12-02
> This is pretty cool, but I wonder if this should be opt-in given that the performance advantage is fairly negligible unless you are doing a lot digesting vs the breadth of code that relies on sha256.

Are you concerned mostly with bugs in this implementation or reliability of the CPU instructions or something else?

> This could even be a plugin that is disabled by default where the plugin init would register the hasher.

Another alternative would be to make this a build tag.  Then there wouldn't be an extra plugin hanging around if this gets folded into Go.