@AkihiroSuda 2022-12-08
> Does the Go upstream have aarch64 support already? 

Yes https://github.com/golang/go/blob/master/src/crypto/sha256/sha256block_arm64.s